Created module consensus-utility. Moved EventConfig and TransactionConfig there.
