+++ platform-sdk/consensus-utility/build.gradle.kts
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: Apache-2.0
+plugins {
+    id("org.hiero.gradle.module.library")
+    id("org.hiero.gradle.feature.publish-artifactregistry")
+    id("org.hiero.gradle.feature.test-fixtures")
+}
+
+description = "Consensus Utility"
+
+mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
