+++ platform-sdk/consensus-utility/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+module org.hiero.consensus.utility {
+    exports org.hiero.consensus.config;
+
+    requires transitive com.swirlds.config.api;
+    requires transitive org.hiero.consensus.model;
+    requires static transitive com.github.spotbugs.annotations;
+}
