+++ platform-sdk/swirlds-common/src/testFixtures/java/module-info.java
@@ -27,6 +27,7 @@ open module com.swirlds.common.test.fixtures {
     requires com.swirlds.logging;
     requires com.swirlds.platform.core;
     requires org.hiero.base.utility.test.fixtures;
+    requires org.hiero.consensus.utility;
     requires lazysodium.java;
     requires org.apache.logging.log4j.core;
     requires org.apache.logging.log4j;
