+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/stale/DefaultStaleEventDetector.java
@@ -3,12 +3,12 @@ package com.swirlds.platform.event.stale;
 
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.component.framework.transformers.RoutableData;
-import com.swirlds.platform.eventhandling.EventConfig;
 import com.swirlds.platform.sequence.map.StandardSequenceMap;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import org.hiero.consensus.config.EventConfig;
 import org.hiero.consensus.model.event.AncientMode;
 import org.hiero.consensus.model.event.EventDescriptorWrapper;
 import org.hiero.consensus.model.event.PlatformEvent;
