+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/consensus/ConsensusTestArgs.java
@@ -12,9 +12,9 @@ import static com.swirlds.common.test.fixtures.WeightGenerators.SINGLE_NODE_STRO
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.test.fixtures.platform.TestPlatformContextBuilder;
 import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
-import com.swirlds.platform.eventhandling.EventConfig_;
 import com.swirlds.platform.test.fixtures.consensus.ConsensusTestParams;
 import java.util.stream.Stream;
+import org.hiero.consensus.config.EventConfig_;
 
 public class ConsensusTestArgs {
