+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/pool/TransactionPoolNexusTest.java
@@ -13,9 +13,9 @@ import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.metrics.noop.NoOpMetrics;
 import com.swirlds.common.test.fixtures.Randotron;
 import com.swirlds.config.api.Configuration;
-import com.swirlds.platform.config.TransactionConfig;
 import java.time.Duration;
 import java.util.List;
+import org.hiero.consensus.config.TransactionConfig;
 import org.hiero.consensus.event.creator.impl.config.EventCreationConfig;
 import org.hiero.consensus.model.status.PlatformStatus;
 import org.junit.jupiter.api.BeforeEach;
