+++ platform-sdk/swirlds-platform-core/src/testFixtures/java/module-info.java
@@ -22,6 +22,7 @@ module com.swirlds.platform.core.test.fixtures {
     requires com.swirlds.merkledb;
     requires com.swirlds.state.api.test.fixtures;
     requires org.hiero.base.utility.test.fixtures;
+    requires org.hiero.consensus.utility;
     requires com.github.spotbugs.annotations;
     requires com.google.common;
     requires java.desktop;
